mountain_pass/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
drf-nested-routers==0.93.4
drf-writable-nested==0.7.0
Markdown==3.4.4
Pillow==10.0.0
psycopg2==2.9.7
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3



